// src/lib/services/models/StandardResponseOfboolean.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
export type StandardResponseOfboolean = {
    success?: boolean;
    message?: string | null;
    data?: boolean;
    statusCode?: HttpStatusCode;
    errors?: any;
    timestamp?: string;
};